refactor: 네이밍 대소문자 수정
