+++ src/main/java/chess/strategy/BishopMoveStrategy.java
@@ -1,4 +1,4 @@
-package chess.Strategy;
+package chess.strategy;
 
 import chess.domain.board.Horizontal;
 import chess.domain.board.Position;
@@ -6,38 +6,11 @@ import chess.domain.board.Vertical;
 import chess.domain.piece.Direction;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
-public class QueenMoveStrategy extends LinearMoveStrategy {
+public class BishopMoveStrategy extends LinearMoveStrategy {
     @Override
     public List<Position> searchMovablePositions(Position target) {
-        List<Position> movablePositions = new ArrayList<>();
-
-        movablePositions.addAll(rookPosition(target));
-        movablePositions.addAll(bishopPosition(target));
-
-        return movablePositions;
-    }
-
-    private List<Position> rookPosition(Position target) {
-        List<Position> movablePositions = new ArrayList<>();
-        Vertical targetVertical = target.getVertical();
-        Horizontal targetHorizontal = target.getHorizontal();
-
-        for (Horizontal horizontal : Horizontal.values()) {
-            movablePositions.add(Position.of(horizontal, targetVertical));
-        }
-        for (Vertical vertical : Vertical.values()) {
-            movablePositions.add(Position.of(targetHorizontal, vertical));
-        }
-
-        movablePositions.removeAll(Collections.singletonList(target));
-
-        return movablePositions;
-    }
-
-    private List<Position> bishopPosition(Position target) {
         List<Position> movablePositions = new ArrayList<>();
 
         movablePositions.addAll(calculateBishopMovablePositions(target, Direction.RIGHT_TOP));
